[MOVE] Add location folder and move location files
